pyproject.toml
[project]
name = "repo-radar"
version = "0.1.1"
description = "GitHub Pull Request audit toolkit with LLM and CLI support"
readme = "README.md"
authors = [{ name = "Karthick Shanmuga Rao", email = "karthikssfeb@gmail.com" }]
requires-python = ">=3.10"
dependencies = [
    "PyGithub>=2.1.1",
    "pydantic>=2.0",
    "fastapi>=0.103.0",
    "uvicorn>=0.23.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.67.1",
    "openai>=1.97.1",
    "fastmcp>=2.10.6",
]

[project.scripts]
repo-radar-config-audit = "repo_radar.audit_scripts.run_config_audit:main"
repo-radar-llm-audit = "repo_radar.audit_scripts.run_llm_audit:main"

[build-system]
requires = ["uv_build>=0.8.3,<0.9.0"]
build-backend = "uv_build"

[tool.uv-build]
package-dir = "src"
